Allow Config to be created without a file.
It's sometimes desirable to create a server that is configured by logic
that creates the `Content`. Prior to this commit, such content needed
to be written to a temporary file and then loaded into memory again in
order to get a complete `Config`.

This commit adds a `FromContent` method, allowing the `Config` to be
created directly from a `Content`.